Sends text message on invitee_created webhook
I moved the `post` outside of the before block. Although it was nice for readability, including the post in the before block meant that I could not set any specific expectations (`expect`) in subsequent tests (since the `post` in the `before` block executes before the expect statement in the `it` block).